Upgrade ruby for new Heroku stack
This app is so old that it's still running on an old Heroku stack
that's about to go away.  In order to run on a new stack we have
to upgrade to a Ruby version that is available on the new stack,
which means we also have to update some dependencies.